Move the ceno update extension into the ceno settings extension.
Rename the extension from ceno settings to ceno.
Change the UI for the settings page to be similar to the Firefox settings UI.
Add placeholders to the settings page for the CENO and Ouinet version numbers.
Move the settings page from the extensions options page to a page embedded in the extension,
as it looks like option pages are not supported for system extesions.
Remove some properties from the manifest that are not supported on Android.