Back to forcing download of log file (with a better name).
Fennec does not offer a clean way of saving a loaded text file as is to
storage (only the immensely useful *Print* and *Save as PDF* options exist),
so just force downloading the file, which should be ok on the desktop too.

However, use a better name with a proper extension to increase the chances of
being able to open in with the right app.