Use less technical texts in settings and popup pages.
They both use the same strings (except for the local cache in the popup).

Add a warning to remind the user that options are not stored.  Keep the link
to the manual to learn what the options do in Ouinet terms.

Also do not group by browsing mode since dynamic warnings offer enough
information.

The Proxy and Injector toggles may be merged together into "Via the CENO
network", but that would require extra JS trickery.