Release v1.2.0.
The main news for this release is the addition of i18n mechanisms allowing for
the localization (l10n) and translation of the Extension.  A sample
translation into Spanish is included, and right-to-left scripts should
hopefully work too (though no translations are included for the moment).

Please check [Weblate][ceno-weblate] if you want to contribute translations!

[ceno-weblate]: https://hosted.weblate.org/projects/censorship-no/ceno-ext-settings/

Features:

  - Support for localization and translation of the Extension (Spanish
    translation included).

Enhancements:

  - Better wording of warnings in settings page.

Bugfixes:

  - Avoid calling Google for fonts when the update page is shown.
  - Avoid unsafe use of `innerHTML` in JS code.