Remove URL fragment when computing DHT group.
Otherwise a client looking for a URL with a fragment would not be able to
retrieve the resource from another client which accessed the same URL without
a fragment (or with a different one), and vice versa.

Also, fragments may contain private data to be handled exclusively at the
agent/browser, like encryption keys.