Release v1.7.0
The main new feature of this release is the support for native messaging
which allows a port between the extension and the application to be
opened and data exchange to occurr between them.
This is also the first version that fully deprecates the use of the
settings and popup html pages in the Android app.

Other minor improvements include the removal of unsupported locales.