Release v1.8.0
Minor update to native messaging API, the JSON object sent by the
extension now includes the URL associated with the tabId. This allows
the source counts to be more easily identified when they are received
on the application side. It also removes the unused sources.html and
sources.js files, which were an experimental solution.