Refactor Sources popup to include public/personal toggle
Removes the public/personal toggle from the settings page and makes the
popup page the main entry point for the extension. Removes the page
action button and uses a single browser action to open the popup, the
settings page will now be accessed through a button on in the popup.