Force personal browsing mode when using private tab
The browsing mode buttons will now be disabled and show the selected
mode as personal when the active tab in the current window has the
incognito flag set true. Additionally, fixes were made to correct the
behavior of the source counts, so they update while the popup is open
and so they reflect the counts for the active tab in the current window.