Add fixes to make desktop-refactor compatible w/ Android
The geckoViewAddons and nativeMessagingFromContent were needed for
Android's usage of native messaging. Also, only attempt to add listener
for browser.windows, if it is defined, as this caused crash on Android.
Finally, only set up native messaging if the OS is Android, to avoid
error when extension is used on desktop.